refactor: moved components, service and model files to right folder #trench-102
