+++ frontend/src/app/configuration/measurement-history/components/measurement-history.component.html
