+++ frontend/src/app/configuration/measurement-history/interfaces/measurement.model.ts
