+++ frontend/src/app/configuration/measurement-history/services/measurement-history.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
-
 import { MeasurementHistoryService } from './measurement-history.service';
+
 
 describe('MeasurementHistoryService', () => {
   let service: MeasurementHistoryService;
